Move dialog, sanitizer, and iframe to common
