+++ src/common/dialog.css
