+++ src/docregistry/kernelactions.ts
@@ -7,7 +7,7 @@ import {
 
 import {
   showDialog, cancelButton, warnButton
-} from '../dialog';
+} from '../common/dialog';
 
 
 /**
