+++ src/filebrowser/crumbs.ts
@@ -31,7 +31,7 @@ import {
 
 import {
   showDialog
-} from '../dialog';
+} from '../common/dialog';
 
 import {
   FileBrowserModel
