+++ src/filebrowser/dialogs.ts
@@ -15,7 +15,7 @@ import {
 
 import {
   showDialog
-} from '../dialog';
+} from '../common/dialog';
 
 import {
   DocumentManager
